improved code organization
